requirements.txt
# AI-Powered Resume Analyzer — Python Dependencies
# Install with: pip install -r requirements.txt

# OpenAI SDK
openai==1.54.3

# PDF parsing
pypdf==5.1.0

# Web UI
streamlit==1.40.1

# Environment variable management
python-dotenv==1.0.1
